added res_ attributes in Fields.h5

+++ src/PicParams.h
@@ -142,7 +142,7 @@ public:
     /*! \brief Time resolution.
       Number of timesteps in \f$ 2\pi/\omega_N \f$ where \f$ \omega_N \f$ is the normalization (plasma or laser) frequency
     */
-    unsigned int res_time;
+    double res_time;
 
     //! simulation exit time in units of \f$ 2\pi/\omega_N \f$
     double sim_time;
@@ -150,7 +150,7 @@ public:
     /*! \brief Space resolution.
       Number of cells in every direction in \f$ 2\pi/k_N \f$ where \f$ k_N=\omega_N/c \f$ is the normalization wavenumber
     */
-    std::vector<unsigned int> res_space;
+    std::vector<double> res_space;
 
     //! local simulation box size in \f$2\pi/k_N \f$
     std::vector<double> sim_length;
